Remove long imports from compatibility

+++ sympy/core/numbers.py
@@ -14,7 +14,7 @@ from .evalf import pure_complex
 from .decorators import _sympifyit
 from .cache import cacheit, clear_cache
 from .logic import fuzzy_not
-from sympy.core.compatibility import (as_int, long, HAS_GMPY, SYMPY_INTS,
+from sympy.core.compatibility import (as_int, HAS_GMPY, SYMPY_INTS,
     int_info, gmpy)
 from sympy.core.cache import lru_cache
 
@@ -1175,7 +1175,7 @@ class Float(Number):
                     if not all((
                             num[0] in (0, 1),
                             num[1] >= 0,
-                            all(type(i) in (long, int) for i in num)
+                            all(type(i) in (int, int) for i in num)
                             )):
                         raise ValueError('malformed mpf: %s' % (num,))
                     # don't compute number or else it may
@@ -3885,13 +3885,13 @@ converter[fractions.Fraction] = sympify_fractions
 
 if HAS_GMPY:
     def sympify_mpz(x):
-        return Integer(long(x))
+        return Integer(int(x))
 
     # XXX: The sympify_mpq function here was never used because it is
     # overridden by the other sympify_mpq function below. Maybe it should just
     # be removed or maybe it should be used for something...
     def sympify_mpq(x):
-        return Rational(long(x.numerator), long(x.denominator))
+        return Rational(int(x.numerator), int(x.denominator))
 
     converter[type(gmpy.mpz(1))] = sympify_mpz
     converter[type(gmpy.mpq(1, 2))] = sympify_mpq
